A:サーバー側実績解除の処理を追加
