+++ backend/src/presentation/routes/achievementController.ts
@@ -1,7 +1,10 @@
 import express from "express";
 import { AchievementApplicationService } from "../../application/achievement/achievementApplicationService";
+import { ContributionApplicationService } from "../../application/contributions/contributionApplicationService";
+import { UnlockAchievementApplicationService } from "../../application/unlockAchievement/unlokAchievementApplicationService";
 import { AchievementId } from "../../domain/models/achievement/AchievementId";
 import { AchievementRepository } from "../../domain/models/achievement/achievementRepository";
+import { UnlockAchievementRepository } from "../../domain/models/unlockAchievement/unlockAchievementRepository";
 
 export const router = express.Router();
 
@@ -61,4 +64,15 @@ router
       new AchievementId(req.params.id)
     );
     res.json({ message: "実績の削除に成功しました" });
+  })
+  .post("/unlock", async (req, res) => {
+    const unlockAchievementApplicationService =
+      new UnlockAchievementApplicationService(
+        new UnlockAchievementRepository()
+      );
+    const material =
+      await new ContributionApplicationService().getContributionsMaterial();
+    // @@ エンジニアIDを渡せるようにする
+    unlockAchievementApplicationService.saveFullfilled(1, material);
+    res.json();
   });
